refactor(engine): simplify ssrf-guard and rename folder to network
- Move engine/src/lib/ssrf/ to engine/src/lib/network/ (test path mirrored).
- Throw EngineGenericError instead of a bespoke SSRFBlockedError class;
  engine CLAUDE.md requires ExecutionError subclasses. Error name is kept
  as 'SSRFBlockedError' so existing err.name checks still work.
- Inline namespace methods per server STYLE.md §2 (drop install/uninstall
  alias layer), collapse one-call-site helpers, tidy types + file order.
