+++ packages/server/engine/test/network/ssrf-guard.test.ts
@@ -1,8 +1,9 @@
 import dns from 'node:dns'
 import { Socket, createServer, Server } from 'node:net'
+import { EngineGenericError } from '@activepieces/shared'
 import { EnvHttpProxyAgent, getGlobalDispatcher, setGlobalDispatcher } from 'undici'
 import { afterAll, afterEach, beforeAll, beforeEach, describe, expect, it } from 'vitest'
-import { ssrfGuard, SSRFBlockedError } from '../../src/lib/ssrf/ssrf-guard'
+import { ssrfGuard } from '../../src/lib/network/ssrf-guard'
 
 function connectOnce(options: { host: string, port: number }): Promise<{ connected: boolean, error?: Error }> {
     return new Promise((resolve) => {
@@ -69,7 +70,7 @@ describe('ssrf-guard', () => {
     describe('dns.lookup hook', () => {
         it('rejects hostname resolving to loopback via promises api', async () => {
             ssrfGuard.install({ enabled: true, allowList: [] })
-            await expect(dns.promises.lookup('localhost')).rejects.toBeInstanceOf(SSRFBlockedError)
+            await expect(dns.promises.lookup('localhost')).rejects.toBeInstanceOf(EngineGenericError)
         })
 
         it('rejects hostname resolving to loopback via callback api', async () => {
@@ -77,7 +78,7 @@ describe('ssrf-guard', () => {
             const err = await new Promise<unknown>((resolve) => {
                 dns.lookup('localhost', (e) => resolve(e))
             })
-            expect(err).toBeInstanceOf(SSRFBlockedError)
+            expect(err).toBeInstanceOf(EngineGenericError)
         })
 
         it('honours allowList for normally blocked IP', async () => {
@@ -105,14 +106,14 @@ describe('ssrf-guard', () => {
             ssrfGuard.install({ enabled: true, allowList: [] })
             const result = await connectOnce({ host: '169.254.169.254', port: 80 })
             expect(result.connected).toBe(false)
-            expect(result.error).toBeInstanceOf(SSRFBlockedError)
+            expect(result.error).toBeInstanceOf(EngineGenericError)
         })
 
         it('blocks connect({ host, port }) to RFC1918 IP', async () => {
             ssrfGuard.install({ enabled: true, allowList: [] })
             const result = await connectOnce({ host: '10.0.0.5', port: 443 })
             expect(result.connected).toBe(false)
-            expect(result.error).toBeInstanceOf(SSRFBlockedError)
+            expect(result.error).toBeInstanceOf(EngineGenericError)
         })
 
         it('allows loopback target on a whitelisted port (engine↔worker RPC, egress proxy)', async () => {
@@ -125,7 +126,7 @@ describe('ssrf-guard', () => {
             ssrfGuard.install({ enabled: true, allowList: [], allowedLoopbackPorts: [publicPort + 1] })
             const result = await connectOnce({ host: '127.0.0.1', port: publicPort })
             expect(result.connected).toBe(false)
-            expect(result.error).toBeInstanceOf(SSRFBlockedError)
+            expect(result.error).toBeInstanceOf(EngineGenericError)
         })
 
         it('allowList overrides block for raw IP connect', async () => {
