+++ packages/server/worker/test/e2e/helpers/bundle-ssrf-guard.ts
@@ -2,7 +2,7 @@ import path from 'node:path'
 import { build } from 'esbuild'
 
 export async function bundleSsrfGuard({ outfile }: { outfile: string }): Promise<void> {
-    const entryPoint = path.resolve(__dirname, '../../../../engine/src/lib/ssrf/ssrf-guard.ts')
+    const entryPoint = path.resolve(__dirname, '../../../../engine/src/lib/network/ssrf-guard.ts')
     const sharedSrc = path.resolve(__dirname, '../../../../../shared/src')
     await build({
         entryPoints: [entryPoint],
